Use the "Has Typeclass" approach to make testing easier.
In the previous commit, we saw how the "Has Typeclass" approach allowed
us to de-couple our computations from our `Env` core datatype. Instead
of having computations that know everything about their environment , we
have computations that care about specific values in their environment.

Now that our code is de-coupled, we can test our computations AND only
provide the environment that that computation cares about. Notice that
in the test for `computeGreeting`, we provided only a name as the
environment, since that is all that `computeGreeting` has knowledge of /
cares about.