Regain purity by extracting effects into a typeclass.
Previously, our `program` computation had a `MonadIO` constraint. This
meant that any computation defined in terms of `program` would also be
required to have the `MonadIO` constraint, allowing arbitrary IO in that
"other" computation.

By extracting the effect of "outputting content" to a typeclass, we have
removed the `MonadIO` constraint from `program`. Computations defined in
terms of `program` will no longer have the `MonadIO` constraint and will
therefore not be able to perform arbitrary IO.

Furthermore, by extracting the effect into a typeclass, the
implementation of "outputting content" can vary based on the datatype in
question.