Test IO in a pure way.
Extracting effects - a.k.a. IO, in this context - allowed us to regain
purity (see previous code and commit message in previous commit).

In addition to regaining purity (or, as a result of regaining purity),
we are able to test our effects in a pure way. We simply use a different
monadic stack.

The benefits of creating interfaces and programming to an interface - as
opposed to using a concrete implementation - is given further treatment
in [Chapter
15](http://book.realworldhaskell.org/read/programming-with-monads.html#id649416)
of Real World Haskell.